fixed transactionRollback.

+++ Siege.Provisions/Siege.Provisions/Finders/IQuery.cs
@@ -21,6 +21,6 @@ namespace Siege.Provisions.Finders
 	{
 		IList<T> Find();
 		T FindFirstOrDefault();
-		long Count();
+		int Count();
 	}
 }
